Rename filter by provider view to filter
