+++ gui/src/renderer/components/Filter.tsx
